Extract HMR template to separate file
The template is loaded using esbuild's `file` loader.
